backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-rest-auth==5.0.2
django-allauth==0.61.1
Pillow==10.2.0